Second pass at PR comments.
