+++ static/login.html
@@ -13,6 +13,6 @@
 
 <body>
     <div id="root" ></div>
-    <script type="text/javascript" src="../static/setup-bundle.min.js"></script>
+    <script type="text/javascript" src="../static/login-bundle.min.js"></script>
 </body>
 </html>
