+++ webpack.config.web.js
@@ -11,7 +11,7 @@ const config = {
     entry: {
         'web': './app/index',
         'oauth': './app/oauth/index',
-        'setup': './app/setup/index'
+        'login': './app/login/index'
     },
 
     output: {
